+++ src/NCronJob/Configuration/DependencyBuilder.cs
@@ -1,3 +1,5 @@
+using Microsoft.Extensions.DependencyInjection;
+
 namespace NCronJob;
 
 /// <summary>
@@ -7,6 +9,9 @@ public sealed class DependencyBuilder<TPrincipalJob>
     where TPrincipalJob : IJob
 {
     private readonly List<JobDefinition> dependentJobOptions = [];
+    private readonly IServiceCollection services;
+
+    internal DependencyBuilder(IServiceCollection services) => this.services = services;
 
     /// <summary>
     /// Adds a job that runs after the principal job has finished with a given <paramref name="parameter"/>.
@@ -18,6 +23,24 @@ public sealed class DependencyBuilder<TPrincipalJob>
         where TJob : IJob
     {
         dependentJobOptions.Add(new JobDefinition(typeof(TJob), parameter, null, null));
+        return this;
+    }
+
+    /// <summary>
+    /// Adds an anonymous delegate job that runs after the principal job has finished.
+    /// </summary>
+    /// <param name="jobDelegate">The delegate that represents the job to be executed. This delegate must return either void or Task.</param>
+    public DependencyBuilder<TPrincipalJob> RunJob(Delegate jobDelegate)
+    {
+        ArgumentNullException.ThrowIfNull(jobDelegate);
+
+        var jobPolicyMetadata = new JobExecutionAttributes(jobDelegate);
+        var entry = new JobDefinition(typeof(DynamicJobFactory), null, null, null,
+            JobName: DynamicJobNameGenerator.GenerateJobName(jobDelegate),
+            JobPolicyMetadata: jobPolicyMetadata);
+        dependentJobOptions.Add(entry);
+        services.AddSingleton(entry);
+        services.AddSingleton(new DynamicJobRegistration(entry, sp => new DynamicJobFactory(sp, jobDelegate)));
         return this;
     }